Use signed PitchCompute accumulator

+++ src/RedSound/RedExecute.cpp
@@ -507,7 +507,7 @@ u8 GetRandomData()
 int PitchCompute(int basePitch, int pitchOffset, int wavePitch, int fineTune)
 {
     int value;
-    u32 pitch;
+    int pitch;
     int octaveAdjust;
     int noteBand;
 
@@ -515,7 +515,7 @@ int PitchCompute(int basePitch, int pitchOffset, int wavePitch, int fineTune)
     basePitch >>= REDSOUND_FIXED_SHIFT;
     pitch = pitchOffset + (wavePitch >> 16);
     pitch = basePitch + pitch;
-    while ((int)pitch < 0) {
+    while (pitch < 0) {
         pitch += REDSOUND_PITCH_OCTAVE_UNITS;
         octaveAdjust -= 1;
     }
